Implement schema merging for union-able sync merging
