+++ src/sync/utils.rs
@@ -1,16 +1,15 @@
-use crate::sync::schema::{SyncColumn, SyncSchema};
-use crate::sync::writer::{DataSyncItem, LOWER_SYNC, UPPER_SYNC};
+use crate::sync::schema::SyncSchema;
+use crate::sync::writer::DataSyncItem;
 use crate::sync::SyncResult;
 use arrow::array::{new_null_array, Array, ArrayRef, RecordBatch, Scalar, UInt64Array};
 use arrow::compute::kernels::cmp::{gt_eq, lt_eq};
 use arrow::compute::{and_kleene, bool_or, concat_batches, filter, is_not_null, take};
 use arrow_row::{Row, RowConverter, SortField};
-use clade::sync::{ColumnDescriptor, ColumnRole};
+use clade::sync::ColumnRole;
 use datafusion::functions_aggregate::min_max::{MaxAccumulator, MinAccumulator};
 use datafusion::physical_optimizer::pruning::PruningStatistics;
-use datafusion_common::{Column, DataFusionError, Result, ScalarValue};
-use datafusion_expr::{col, lit, when, Accumulator, Expr};
-use indexmap::IndexSet;
+use datafusion_common::{DataFusionError, Result};
+use datafusion_expr::{col, lit, Accumulator, Expr};
 use std::collections::{HashMap, HashSet, VecDeque};
 use tracing::log::warn;
 
@@ -409,287 +408,6 @@ pub(super) fn get_prune_map(
     }
 
     Ok(prune_map)
-}
-
-// Build the merged projection and column descriptors from the sync schemas
-pub(super) fn merge_schemas(
-    lower_schema: &SyncSchema,
-    upper_schema: &SyncSchema,
-) -> SyncResult<(Vec<ColumnDescriptor>, Vec<Expr>)> {
-    let mut projection = vec![];
-    let mut col_desc = IndexSet::new();
-
-    // Project all sync columns from the lower schema
-    for lower_sync_col in lower_schema.columns() {
-        // Build the expression to project the combined column
-        let expr = match lower_sync_col.role() {
-            ColumnRole::OldPk => project_old_pk_column(lower_sync_col, upper_schema)?,
-            ColumnRole::NewPk => project_new_pk_column(lower_sync_col, upper_schema)?,
-            ColumnRole::Changed => {
-                project_changed_column(lower_sync_col, upper_schema, SyncPosition::Lower)?
-            }
-            ColumnRole::Value => project_value_column(
-                lower_sync_col,
-                lower_schema,
-                upper_schema,
-                &mut col_desc,
-                &mut projection,
-                SyncPosition::Lower,
-            )?,
-        };
-
-        col_desc.insert(lower_sync_col.column_descriptor());
-        projection.push(expr);
-    }
-
-    // Now project any missing sync columns from the upper schema
-    for upper_sync_col in upper_schema.columns() {
-        let cd = upper_sync_col.column_descriptor();
-
-        if col_desc.contains(&cd) {
-            // We've already projected this column from the lower schema
-            continue;
-        }
-
-        // Build the expression that is going to be used when upper sync column value is missing
-        let expr = match upper_sync_col.role() {
-            ColumnRole::Changed => {
-                project_changed_column(upper_sync_col, lower_schema, SyncPosition::Upper)?
-            }
-            ColumnRole::Value => project_value_column(
-                upper_sync_col,
-                upper_schema,
-                lower_schema,
-                &mut col_desc,
-                &mut projection,
-                SyncPosition::Upper,
-            )?,
-            _ => continue, // We already picked up PKs from the lower sync
-        };
-
-        col_desc.insert(upper_sync_col.column_descriptor());
-        projection.push(expr);
-    }
-
-    let col_desc = col_desc.into_iter().collect::<Vec<_>>();
-
-    // Finally use a qualified alias for the projection, which will act as a new lower sync in the
-    // subsequent sync merge
-    let projection = col_desc
-        .iter()
-        .zip(projection)
-        .map(|(cd, expr)| {
-            expr.alias_qualified(Some(LOWER_SYNC), SyncColumn::canonical_field_name(cd))
-        })
-        .collect();
-
-    Ok((col_desc, projection))
-}
-
-enum SyncPosition {
-    Lower,
-    Upper,
-}
-
-fn project_old_pk_column(
-    lower_old_pk_col: &SyncColumn,
-    upper_schema: &SyncSchema,
-) -> SyncResult<Expr> {
-    let lower_old_pk_field = lower_old_pk_col.field().name();
-
-    let upper_old_pk_col = upper_schema
-        .column(lower_old_pk_col.name(), ColumnRole::OldPk)
-        .expect("Old Pk columns must be the same across syncs");
-
-    // For un-matched lower sync rows use lower sync old PKs, otherwise
-    // for un-matched upper sync rows use upper sync old PKs, otherwise
-    // for matched lower-upper sync rows use lower sync old PKs
-    Ok(when(
-        col(UPPER_SYNC).is_null(),
-        col(Column::new(Some(LOWER_SYNC), lower_old_pk_field)),
-    )
-    .otherwise(
-        when(
-            col(LOWER_SYNC).is_null(),
-            col(Column::new(
-                Some(UPPER_SYNC),
-                upper_old_pk_col.field().name(),
-            )),
-        )
-        .otherwise(col(Column::new(Some(LOWER_SYNC), lower_old_pk_field)))?,
-    )?)
-}
-
-fn project_new_pk_column(
-    lower_new_pk_col: &SyncColumn,
-    upper_schema: &SyncSchema,
-) -> SyncResult<Expr> {
-    let lower_new_pk_field = lower_new_pk_col.field().name();
-
-    let upper_new_pk_col = upper_schema
-        .column(lower_new_pk_col.name(), ColumnRole::NewPk)
-        .expect("Old Pk columns must be the same across syncs");
-
-    // For un-matched lower sync rows use lower sync new PKs, otherwise
-    // for un-matched and matched upper sync rows use upper sync new PKs
-    Ok(when(
-        col(UPPER_SYNC).is_null(),
-        col(Column::new(Some(LOWER_SYNC), lower_new_pk_field)),
-    )
-    .otherwise(col(Column::new(
-        Some(UPPER_SYNC),
-        upper_new_pk_col.field().name(),
-    )))?)
-}
-
-fn project_changed_column(
-    this_changed_col: &SyncColumn,
-    other_schema: &SyncSchema,
-    sync_position: SyncPosition,
-) -> SyncResult<Expr> {
-    let (this_sync, other_sync) = match sync_position {
-        SyncPosition::Lower => (LOWER_SYNC, UPPER_SYNC),
-        SyncPosition::Upper => (UPPER_SYNC, LOWER_SYNC),
-    };
-
-    // For un-matched rows in this sync use this sync changed values, otherwise ...
-    Ok(when(
-        col(other_sync).is_null(),
-        col(Column::new(
-            Some(this_sync),
-            this_changed_col.field().name(),
-        )),
-    )
-    .otherwise(
-        if let Some(other_changed_col) =
-            other_schema.column(this_changed_col.name(), ColumnRole::Changed)
-        {
-            // ... other sync has the corresponding `Changed` column, take the upper column value in
-            // both cases
-            col(Column::new(
-                Some(UPPER_SYNC),
-                match sync_position {
-                    SyncPosition::Lower => other_changed_col.field().name(),
-                    SyncPosition::Upper => this_changed_col.field().name(),
-                },
-            ))
-        } else if other_schema
-            .column(this_changed_col.name(), ColumnRole::Value)
-            .is_some()
-        {
-            // ... other sync doesn't have the `Changed` column in the schema, but it does have
-            // the corresponding `Value` column
-            match sync_position {
-                // project true to pick those upper values up
-                SyncPosition::Lower => lit(true),
-                // project true to pick those lower values up only for non-upper rows
-                SyncPosition::Upper => {
-                    when(col(UPPER_SYNC).is_null(), lit(true)).otherwise(col(
-                        Column::new(Some(UPPER_SYNC), this_changed_col.field().name()),
-                    ))?
-                }
-            }
-        } else {
-            // ... other sync has neither the `Changed` nor the associated `Value` column,
-            match sync_position {
-                // project false to avoid overriding lower values with NULL values
-                SyncPosition::Lower => lit(false),
-                // project false to avoid overriding upper values with NULL values
-                SyncPosition::Upper => {
-                    when(col(UPPER_SYNC).is_null(), lit(false)).otherwise(col(
-                        Column::new(Some(UPPER_SYNC), this_changed_col.field().name()),
-                    ))?
-                }
-            }
-        },
-    )?)
-}
-
-fn project_value_column(
-    this_value_col: &SyncColumn,
-    this_schema: &SyncSchema,
-    other_schema: &SyncSchema,
-    col_desc: &mut IndexSet<ColumnDescriptor>,
-    projection: &mut Vec<Expr>,
-    sync_position: SyncPosition,
-) -> SyncResult<Expr> {
-    let (this_sync, other_sync) = match sync_position {
-        SyncPosition::Lower => (LOWER_SYNC, UPPER_SYNC),
-        SyncPosition::Upper => (UPPER_SYNC, LOWER_SYNC),
-    };
-
-    if other_schema
-        .column(this_value_col.name(), ColumnRole::Value)
-        .is_none()
-        && this_schema
-            .column(this_value_col.name(), ColumnRole::Changed)
-            .is_none()
-    {
-        // There's no corresponding `Value` sync column in the other schema (so we'll project NULLs),
-        // and there's no `Changed` column in this schema either, we need to project the `Changed`
-        // column that takes all the values from this sync and no value from the other sync
-        col_desc.insert(ColumnDescriptor {
-            role: ColumnRole::Changed as _,
-            name: this_value_col.name().to_string(),
-        });
-        projection.push(when(col(this_sync).is_null(), lit(false)).otherwise(lit(true))?);
-    }
-
-    // For un-matched rows in this sync use this sync changed values, otherwise ...
-    Ok(when(
-        col(other_sync).is_null(),
-        col(Column::new(Some(this_sync), this_value_col.field().name())),
-    )
-    .otherwise(
-        if let Some(other_value_col) =
-            other_schema.column(this_value_col.name(), ColumnRole::Value)
-        {
-            match sync_position {
-                SyncPosition::Lower => {
-                    // ... the upper sync has this `Value` column too, take its values
-                    col(Column::new(
-                        Some(UPPER_SYNC),
-                        other_value_col.field().name(),
-                    ))
-                }
-                SyncPosition::Upper => {
-                    // ... the lower sync has this `Value` column too, take its values where it
-                    // didn't get joined to the upper sync, otherwise take upper sync values
-                    when(
-                        col(UPPER_SYNC).is_null(),
-                        col(Column::new(Some(UPPER_SYNC), this_value_col.field().name())),
-                    )
-                    .otherwise(col(Column::new(
-                        Some(LOWER_SYNC),
-                        other_value_col.field().name(),
-                    )))?
-                }
-            }
-        } else {
-            // ... other sync doesn't have this column but this sync does. Since the corresponding
-            // `Changed` column will be added regardless
-            match sync_position {
-                SyncPosition::Lower => {
-                    // ... project NULLs in its place for unmatched upper sync rows, otherwise for
-                    // the matched rows project the lower sync values.
-                    when(
-                        col(UPPER_SYNC).is_null(),
-                        lit(ScalarValue::Null
-                            .cast_to(this_value_col.field().data_type())?),
-                    )
-                    .otherwise(col(Column::new(
-                        Some(LOWER_SYNC),
-                        this_value_col.field().name(),
-                    )))?
-                }
-                SyncPosition::Upper => {
-                    // ... project upper schema column values both where it matched rows in the
-                    // lower sync and where it didn't (i.e. which has NULLs by virtue of the JOIN)
-                    col(Column::new(Some(UPPER_SYNC), this_value_col.field().name()))
-                }
-            }
-        },
-    )?)
 }
 
 #[cfg(test)]
